support for multi s3
