+++ packages/discovery-provider/plugins/pedalboard/apps/mri/src/index.ts
@@ -3,7 +3,7 @@ import { S3Client } from "@aws-sdk/client-s3"
 import { logger } from "./logger"
 import { CronJob } from "cron"
 import { webServer } from "./server"
-import { createBucket, listBuckets } from "./s3"
+import { createBucket } from "./s3"
 import { clm } from "./clm"
 
 const main = async () => {
@@ -14,7 +14,7 @@ const main = async () => {
         connection: process.env.audius_db_url || "postgresql://postgres:pass@0.0.0.0:5433/default_db"
     })
 
-    const s3 = new S3Client({
+    const localstackS3 = new S3Client({
         region: "us-east-1",
         endpoint: "http://localhost:4566",
         credentials: {
@@ -23,22 +23,29 @@ const main = async () => {
         },
         forcePathStyle: true
     })
+    const localstackS3Config = {
+        s3: localstackS3,
+        bucket: "audius-clm-data",
+        keyPrefix: "Audius_CLM_"
+    }
 
-    await createBucket("audius-clm-data")
+    await createBucket(localstackS3Config.bucket)
+
+    const s3s = [localstackS3Config]
 
     const job = CronJob.from({
         // run at 10 AM PST every day
         cronTime: '00 00 10 * * *',
         onTick: async function () {
             const date = new Date()
-            await clm(db, s3, date)
+            await clm(db, s3s, date)
         },
         timeZone: "America/Los_Angeles"
     })
 
     job.start()
 
-    const server = webServer(db, s3)
+    const server = webServer(db, s3s)
     const port = process.env.audius_mri_port || 6003
     server.listen(port, () => logger.info({ port }, "webserver is running"))
 }
